@@ -1,5 +1,3 @@
-import userEvent from '@testing-library/user-event';
-
 import Button from '@components/Buttons/Button';
 import { render } from '@tests/test-utils';
 
@@ -8,7 +6,7 @@ describe('Button Component', () => {
     const onClick = jest.fn();
     const { getByRole } = render(<Button onClick={onClick}>Button</Button>);
     const button = getByRole('button');
-    userEvent.click(button);
+    button.click();
     expect(onClick).toBeCalled();
   });
 });
